Add a DockerTest tag and two sbt filters 'docker:test' and 'nodocker:test'
