+++ src/test/scala/cromwell/CromwellSpec.scala
@@ -1,25 +1,7 @@
 package cromwell
 
-import akka.actor.ActorSystem
-import akka.testkit.{DefaultTimeout, EventFilter, ImplicitSender, TestKit}
-import org.scalatest.concurrent.ScalaFutures
-import org.scalatest.{BeforeAndAfterAll, Matchers, WordSpecLike}
+import org.scalatest.Tag
 
-abstract class CromwellSpec(actorSystem: ActorSystem) extends TestKit(actorSystem) with DefaultTimeout
-with ImplicitSender with WordSpecLike with Matchers with BeforeAndAfterAll with ScalaFutures {
-
-  def startingCallsFilter(callNames: String*): EventFilter =
-    EventFilter.info(message = s"Starting calls: ${callNames.mkString(", ")}", occurrences = 1)
-
-  def waitForHandledMessage[T](named: String)(block: => T): T = {
-    waitForHandledMessagePattern(s"^received handled message $named")(block)
-  }
-
-  def waitForHandledMessagePattern[T](pattern: String)(block: => T): T = {
-    EventFilter.debug(pattern=pattern, occurrences = 1).intercept {
-      block
-    }
-  }
-
-  protected def getActorSystem = actorSystem
+object CromwellSpec {
+  object DockerTest extends Tag("DockerTest")
 }
