+++ src/test/scala/cromwell/engine/SingleWorkflowRunnerActorSpec.scala
@@ -5,10 +5,10 @@ import cromwell.util.SampleWdl.ThreeStep
 import akka.actor.ActorSystem
 import akka.testkit.EventFilter
 import com.typesafe.config.ConfigFactory
-import cromwell.CromwellSpec
+import cromwell.CromwellTestkitSpec
 import scala.language.postfixOps
 
-class SingleWorkflowRunnerActorSpec extends CromwellSpec(ActorSystem("ActorWorkflowManagerSpec", ConfigFactory.parseString(Config))) {
+class SingleWorkflowRunnerActorSpec extends CromwellTestkitSpec(ActorSystem("ActorWorkflowManagerSpec", ConfigFactory.parseString(Config))) {
   val actorSystem = super.getActorSystem
   val workflowManagerActor = actorSystem.actorOf(WorkflowManagerActor.props)
   val props = SingleWorkflowRunnerActor.props(ThreeStep.WdlSource, ThreeStep.RawInputs, workflowManagerActor)
